Super license slap
Since all of these apply to public code and the repo was private until then (not to mention these new licenses are more permissive), there shouldn't be any issues with this.

Unless, of course, the legal weirdness I created to allow linking with libraries that may or may not be under the GPL is not actually defensible. I'd contact a lawyer if people were to actually use this, but they won't, so I'll just pass the burden on to anybody who decides to fork this.